Multi-keg-meter Ticking/Updates
Removed “synchronous” waiting animations in the KegLoadMeter class.
Things are now “ticked” and keep track of the amount of delay that has
happened so far and they now animate based on that delay. Drawing and
delay now happen once at the end of the Arduino loop instead of inside
every animate function call.